fix: use double division in success rate calculation
Integer division was causing success rate to always be 0 or 100,
losing decimal precision (e.g., 5/10 = 0 instead of 50.0).

+++ src/main/java/com/wpanther/storage/infrastructure/config/metrics/DocumentStorageMetricsService.java
@@ -263,7 +263,7 @@ public class DocumentStorageMetricsService {
      */
     public double getPdfDownloadSuccessRate() {
         double total = pdfDownloadSuccessCounter.count() + pdfDownloadFailureCounter.count();
-        return total > 0 ? (pdfDownloadSuccessCounter.count() / total) * 100 : 0;
+        return total > 0 ? (pdfDownloadSuccessCounter.count() / (double) total) * 100 : 0;
     }
 
     // ========== Orphaned Document Metrics ==========
